Move logic to write extended file properties part to separate writer. It decreases the clutter and hopefully increases encapsulation, so one writer can optimize how to write its own part without another 4k+ LOC.
